% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{add_datetime}
\alias{add_datetime}
\title{Add Date/Time variables}
\usage{
add_datetime(DT = NULL)
}
\arguments{
\item{DT}{}
}
\value{
LUDT with datetime variables formatted to long-string format.
}
\description{
Expects a data table with the variables EVENT_ID, TZ, YEAR, MONTH_NAME, 
     BEGIN_DAY, BEGIN_DATE_TIME, END_DAY, and END_DATE_TIME. All variables 
     except TZ come in the get_data() call requesting the \emph{details} 
     dataset.
}
\details{
Two new variables are created and added to the data table: BEGIN_DATE and 
     END_DATE. These dates are formatted like %Y-%m-%d %h:%m:%s %z. Then 
     each variable has time readjusted to UTC with lubridate::with_tz(). 
     The new variables are then added to LUDT.
}

