% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{add_fips}
\alias{add_fips}
\title{Add FIPS codes to LUDT}
\usage{
add_fips(DT = NULL)
}
\arguments{
\item{DT}{data table consisting of EVENT_ID, CZ_FIPS, STATE_FIPS}
}
\value{
adds FIPS codes to lookup data table associated w/ EVENT_ID
}
\description{
Add FIPS codes to LUDT
}
\examples{
\dontrun{
add_fips(DT[, .(EVENT_ID, CZ_FIPS, STATE_FIPS)])
}
}

