% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{check_type}
\alias{check_type}
\title{Check if Type is valid character vector}
\usage{
check_type(type = NULL)
}
\arguments{
\item{type}{character, vector}
}
\value{
type
}
\description{
Check if Type is valid character vector
}

