% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{check_year}
\alias{check_year}
\title{Check if Year is numeric input}
\usage{
check_year(year = NULL)
}
\arguments{
\item{numeric, }{vector}
}
\value{
year
}
\description{
Check if Year is numeric input
}

