% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{clean_details}
\alias{clean_details}
\title{clean_details}
\usage{
clean_details(df)
}
\arguments{
\item{df}{Raw `details` dataset.}
}
\value{
Clean dataframe
}
\description{
Clean most variables in the details dataset.
}
\details{
Specifically what is changed:
\itemize{
  \item BEGIN_DATE_TIME, END_DATE_TIME reformatted to "%Y-%m-%d %T", 
    character class with no timezone.
  \item EVENT_TYPE is moved into its own dataframe and cleaned up to match 
    requirements in \href{http://www.nws.noaa.gov/directives/sym/pd01016005curr.pdf}{NWS Instruction 10-1605, March 23, 2016}, 
    Section 2.1.1, pg. 6
  \item DAMAGE_PROPERTY reformatted to numeric values
  \item DAMAGE_CROPS reformatted to numeric values
  \item FLOOD_CAUSE removed spaces between slashes
}

Variables removed:
\itemize{
  \item BEGIN_YEARMONTH
  \item BEGIN_DAY
  \item BEGIN_TIME
  \item END_YEARMONTH
  \item END_DAY
  \item END_TIME
  \item YEAR
  \item MONTH_NAME
  \item CZ_TIMEZONE
  \item EVENT_TYPE (located in `details_event_types`)
}
}

