% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatalities.R
\name{clean_fatalities}
\alias{clean_fatalities}
\title{clean_fatalities}
\usage{
clean_fatalities(df)
}
\arguments{
\item{df}{Raw fatalities dataset}
}
\value{
df[,c("EVENT_ID", "FATALITY_ID", "FATALITY_DATE", "FATALITY_TYPE", "FATALITY_AGE", "FATALITY_SEX", "FATALITY_LOCATION")]
}
\description{
Clean the fatalities dataset
}
\details{
Minor changes. 

FATALITY_DATE is reformatted to '%Y-%m-%d %Y' format. 
FATALITY_LOCATION is cleaned up to match 
\href{NWS Instruction 10-1605, March 23, 2016}{http://www.nws.noaa.gov/directives/sym/pd01016005curr.pdf}
Section 2.6.1.2, p. 11

Varaiables removed:
\itemize{
  \item FAT_YEARMONTH
  \item FAT_DAY
  \item FAT_TIME
  \item EVENT_YEARMONTH
}
}

