% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{clean_locations}
\alias{clean_locations}
\title{clean_locations}
\usage{
clean_locations(df)
}
\arguments{
\item{df}{Raw locations dataset}
}
\value{
df Reorganized raw dataset
}
\description{
"Clean" locations dataset
}
\details{
Nothing is cleaned at the moment; it is just reorganized. EPISODE_ID 
is dropped as it is redundant (in `details`).

I cannot find any documentation on what the purpose of LAT2 and LON2 is 
therefore it is untouched for the moment. No other values require cleaning 
at this time.
}

