% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{cpfz}
\alias{cpfz}
\title{cpfz}
\source{
\url{http://www.nws.noaa.gov/geodata/catalog/wsom/html/cntyzone.htm}
}
\usage{
cpfz()
}
\value{
dataframe
}
\description{
County-Public Forecast Zones Correlation File
}
\details{
This file is an ASCII "dump" in pipe ("|") delimited form, of a "master" 
  shapefile maintained by AGMG for the maintenance of County, Public Forecast 
  Zones, CWA boundaries, and Time zones. Each record represents a single 
  polygon within the master file. Multiple polygons may exist for the same 
  zone/county that contain the same attributes, but the lat and lon fields 
  will be different as they are the centroid of the polygon. No attempt has 
  been made to remove these seemingly duplicate entries from this text file. 
  In addition, the coastal and offshore marine zone shapefiles are also 
  "dumped" to this file.
}

