% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{cpfz_tz_abbr}
\alias{cpfz_tz_abbr}
\title{cpfz_tz_abbr}
\source{
\url{http://www.nws.noaa.gov/geodata/catalog/county/html/county.htm}
}
\usage{
cpfz_tz_abbr()
}
\value{
a list to map cpfz_tz_dt$TIME_ZONE to the true timezone
}
\description{
Return timezone abbreviations for the cpfz dataframe
}
\details{
Two letters appear for the ten (10) counties which are divided by a time 
zone boundary, which are located in the states of AK, FL, ID, ND, NE, OR 
and SD. 
\describe{
  \item{A}{Alaska Standard}
  \item{C}{Central Standard}
  \item{E}{Eastern Standard (e = advanced time not observed)}
  \item{G}{Guam & Marianas}
  \item{H}{Hawaii-Aleutian Standard}
  \item{M}{Mountain Standard (m = advanced time not observed)}
  \item{P}{Pacific Standard}
  \item{S}{Samoa Standard}
  \item{V}{Atlantic Standard}
}
}

