% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{datetime_fields}
\alias{datetime_fields}
\title{datetime_fields}
\usage{
datetime_fields()
}
\value{
a character vector of required date/time variables
}
\description{
Required fields for datetime objects
}
\examples{
  \dontrun{
    if(all(names(DT)) \%in\% datetime_fields())...
    names(DT) <- datetime_fields()
  }
}

