% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details_date_time}
\alias{details_date_time}
\title{details_date_time}
\usage{
details_date_time(df)
}
\arguments{
\item{df[, c("EVENT_ID", }{"[BEGIN|END]_YEARMONTH", "[BEGIN|END]_DAY", "[BEGIN|END]_DATE_TIME")]}
}
\value{
df[,c("EVENT_ID", "[BEGIN|END]_DATE_TIME")] with the datetime value in '%Y-%m-%d %T' format.
}
\description{
Reformat BEGIN_DATE_TIME or END_DATE_TIME to '%Y-%m-%d %T' format.
}
\details{
Takes one of details$BEGIN_DATE_TIME or details$END_DATE_TIME 
along with details$[BEGIN|END]_YEARMONTH, details$[BEGIN|END]_DAY 
and returns a clean, reformatted value of class character with no timezone. 
Times are assumed to be local times.
}

