% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details_event_types}
\alias{details_event_types}
\title{details_event_types}
\usage{
details_event_types(df)
}
\arguments{
\item{df}{A dataframe with EVENT_ID and EVENT_TYPE}
}
\value{
TRUE on success
}
\description{
Extract details$EVENT_TYPE and move to new dataframe.
}
\details{
details$EVENT_TYPE consists mostly of single values. Some `EVENT_ID` 
have multiple `EVENT_TYPE`. These are split into single values and cleaned up 
to match values in 
\href{NWS Instruction 10-1605, March 23, 2016}{http://www.nws.noaa.gov/directives/sym/pd01016005curr.pdf} 
Section 2.1.1, pg. 6.

Because this could potentially duplicate `EVENT_ID` it is kept separate from 
details.
}

