% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details_flood_cause}
\alias{details_flood_cause}
\title{details_flood_cause}
\usage{
details_flood_cause(df)
}
\arguments{
\item{df[, c("EVENT_ID", }{"FLOOD_CAUSE")]}
}
\value{
Complete.cases dataframe with clean FLOOD_CAUSE values.
}
\description{
Clean details$FLOOD_CAUSE
}

