% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{dff_categories}
\alias{dff_categories}
\title{dff_categories}
\format{A 3x5 dataframe
\describe{
  \item{category}{Category}
  \item{max_wind}{Maximum wind in miles per hour}
  \item{damage}{Brief description of typical damage}
}}
\source{
\url{http://www.nhc.noaa.gov/aboutsshws.php}
}
\usage{
dff_categories()
}
\value{
dataframe
}
\description{
Categories and brief descriptions for hurricanes/typhoons.
}

