% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{dff_event_types}
\alias{dff_event_types}
\title{dff_event_types}
\format{A 2x55 dataframe}
\source{
\url{http://www.nws.noaa.gov/directives/sym/pd01016005curr.pdf}{NWS Instruction 10-1605}.
}
\usage{
dff_event_types()
}
\value{
dataframe
}
\description{
Valid Event Types
}

