% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{dff_fatality_locations}
\alias{dff_fatality_locations}
\title{dff_fatality_locations}
\source{
\url{ftp://ftp.ncdc.noaa.gov/pub/data/swdi/stormevents/csvfiles/Storm-Data-Export-Format.docx}
}
\usage{
dff_fatality_locations()
}
\value{
dataframe
}
\description{
Valid fatality locations
}

