% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{dff_tor_f_scales}
\alias{dff_tor_f_scales}
\title{dff_tor_f_scales}
\format{A 6x3 dataframe}
\source{
\url{http://www.spc.noaa.gov/efscale/}
}
\usage{
dff_tor_f_scales()
}
\value{
dataframe
}
\description{
Valid Enhanced Fujita Scale Values
}
\details{
As best as I can tell these values are identical to others that 
exist in the database; i.e., F0, F1, F2...
}

