% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatalities.R
\name{fatalities_fatality_location}
\alias{fatalities_fatality_location}
\title{fatalities_fatality_location}
\usage{
fatalities_fatality_location(df)
}
\arguments{
\item{df[, c("FATALITY_ID", }{"FATALITY_LOCATION")]}
}
\value{
df Clean dataframe
}
\description{
Clean FATALITY_LOCATION to match NWS specifications.
}
\details{
See 
\href{NWS Instruction 10-1605, March 23, 2016}{http://www.nws.noaa.gov/directives/sym/pd01016005curr.pdf}
Section 2.6.1.2, p. 11 for valid fatality locations.
}

