% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{fatality_locations}
\alias{fatality_locations}
\title{fatality_locations}
\usage{
fatality_locations()
}
\value{
dataframe
}
\description{
Dataframe to hold valid fatalities$location per p.8 of 
  \href{ftp://ftp.ncdc.noaa.gov/pub/data/swdi/stormevents/csvfiles/Storm-Data-Export-Format.docx}{Storm Data Export Format document}
}
\details{
Initialize dataframes

These dataframes are generated as reference tables from the original three 
datasets. For example, in details$STATE are set strings of values that sould 
exist. In this file is a function to build a dataframe that holds those keys 
and values.
}

