% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{fips}
\alias{fips}
\title{FIPS URL}
\usage{
fips()
}
\value{
url to fips dataset
}
\description{
This file is an ASCII "dump" in pipe ("|") delimited form, of a "master" 
shapefile maintained by AGMG for the maintenance of County, Public Forecast 
Zones, CWA boundaries, and Time zones. Each record represents a single 
polygon within the master file. Multiple polygons may exist for the same 
zone/county that contain the same attributes, but the lat and lon fields 
will be different as they are the centroid of the polygon. No attempt has 
been made to remove these seemingly duplicate entries from this text file. 
In addition, the coastal and offshore marine zone shapefiles are also 
"dumped" to this file.
}
\examples{

url <- fips()

}

