% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{fips_names}
\alias{fips_names}
\title{Names for fips data table}
\usage{
fips_names()
}
\value{
a character vector, length 10, of names to assign to fips_tz_dt()
}
\description{
Names for fips data table
}
\examples{

setnames(fips_dt) <- fips_names()

}

