% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{fips_tz_abbr}
\alias{fips_tz_abbr}
\title{Return timezone abbreviations for the FIPS dataset}
\usage{
fips_tz_abbr()
}
\value{
a character vector to map fips_tz_dt$TIME_ZONE to the true 
timezone
}
\description{
See \url{http://www.nws.noaa.gov/geodata/catalog/county/html/county.htm}
}
\details{
Two letters appear for the ten (10) counties which are divided by a time 
zone boundary, which are located in the states of AK, FL, ID, ND, NE, OR 
and SD.

\describe{
     \item{A}{Alaska Standard}
     \item{C}{Central Standard}
     \item{E}{Eastern Standard (e = advanced time not observed)}
     \item{G}{Guam & Marianas}
     \item{H}{Hawaii-Aleutian Standard}
     \item{M}{Mountain Standard (m = advanced time not observed)}
     \item{P}{Pacific Standard}
     \item{S}{Samoa Standard}
     \item{V}{Atlantic Standard}
}
}
\examples{
\dontrun{
tz_abbr <- fips_tz_abbr()
}
}

