% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(year = NULL, type = types_available(), clean = TRUE)
}
\arguments{
\item{year}{A numeric vector}

\item{type}{A character vector}

\item{clean}{Clean the dataset. TRUE by default.}
}
\value{
Returns a data.table for each year and type requested.
}
\description{
Download NCDC Storm Events datasets
}
\details{
This is a wrapper function for \code{\link{get_details}}, 
  \code{\link{get_fatalities}} and \code{\link{get_locations}}. This function 
  can load all requested datasets. Strongly recommended to use 
  get_listings() first to see what data is available and the file 
  size of the datasets before downloading datasets.
}
\examples{
  \dontrun{
    get_data(year = c(2015), type = c("details"))
  }
}

