% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_datasets}
\alias{get_datasets}
\title{get_datasets}
\usage{
get_datasets(dt_names, cn, ct)
}
\arguments{
\item{dt_names}{list of Name from get_listings}

\item{cn}{list of column names for Type}

\item{ct}{vector of column types for Type}
}
\value{
a data table of all years requested by Type
}
\description{
get datasets
}

