% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{get_details}
\alias{get_details}
\title{get_details}
\usage{
get_details(year = NULL, clean = TRUE)
}
\arguments{
\item{year}{numeric vector with year(s) to retrieve data.}

\item{clean}{clean the data if TRUE (default). Otherwise, return as-is.}
}
\description{
Download the details datasets for \code{year}.
}
\details{
A warning message will be generated.

  \emph{Too few values at n locations}

  This is acceptable. EVENT_TYPES is rebuilt to capture one event type per 
  row. In most cases there is only one or two event types per event but there 
  can be at most 3 (so far). The message is simply stating there are not 3 
  event types for each row (location). This message is generated by 
  tidyr::separate.
}

