% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatalities.R
\name{get_fatalities}
\alias{get_fatalities}
\title{get_fatalities}
\usage{
get_fatalities(year = NULL, clean = TRUE)
}
\arguments{
\item{year}{numeric vector with year(s) to retrieve data.}

\item{clean}{clean the data if TRUE (default). Otherwise, return as-is.}
}
\value{
dataframe fatalities dataset
}
\description{
Download and, if clean is TRUE, clean fatalities datasets.
}
\details{
In this dataset, FATALITY_ID is the primary key. EVENT_ID is the 
foreign key to JOIN to the details or locations dataset.
}
\examples{
\dontrun{
# Clean fatalities
get_fatalities(2015)
}
\dontrun{
# Raw fatalities
get_fatalities(2015, clean = FALSE)
}
}

