% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_listings}
\alias{get_listings}
\title{Get listing of files from NCDC Storm Events}
\usage{
get_listings()
}
\value{
a data table
}
\description{
This function accesses the NCDC Storm Events dataset listings and returns a 
     formatted data.table showing the StormEvents file names (Name), 
     modified date (Modified), gzip file size (Size), Type (details, 
     fatalities or locations) and Year.
}
\details{
The listings of datasets can be found at 
     \url{http://www1.ncdc.noaa.gov/pub/data/swdi/stormevents/csvfiles/}
}
\examples{
DT <- get_listings()
head(DT, n = 5L)

}

