% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_locations}
\alias{get_locations}
\title{get_locations}
\usage{
get_locations(year = NULL, clean = TRUE)
}
\arguments{
\item{year}{numeric vector of year or years}

\item{clean}{TRUE by default; clean and reorganize dataset.}
}
\value{
dataframe Raw if clean is FALSE, reorganized if TRUE (default)
}
\description{
Download the locations datasets for \code{year}.
}
\details{
This dataset stores some location information of an event/episode.
  There can be multiple events per episode. 
  
  The raw data consists of the following variables in order of appearance 
  with brief descriptions according to the 
  \href{ftp://ftp.ncdc.noaa.gov/pub/data/swdi/stormevents/csvfiles/Storm-Data-Export-Format.docx}{codebook} 
  or, if in italics, my best guess:
  
  \itemize{
    \item YEARMONTH \emph{year and month of event in \%Y\%m format.}
    \item EPISODE_ID ID assigned by NWS to note a storm episode. Has many `EVENT_ID`
    \item EVENT_ID ID assigned by NWS to note an event of an `EPISODE_ID`
    \item LOCATION_INDEX
    \item RANGE
    \item AZIMUTH
    \item LOCATION
    \item LATITUDE
    \item LONGITUDE
    \item LAT2
    \item LON2
  }
}

