% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s.fatalities}
\alias{s.fatalities}
\title{Fatalities Dataset}
\format{A data frame with 11 variables:
\describe{
  \item{FAT_YEARMONTH}{Year and month of the fatality in '\%Y\%m' format}
  \item{FAT_DAY}{Day of the month in '\%e' format}
  \item{FAT_TIME}{Time of the incident in '\%H\%M' format}
  \item{FATALITY_ID}{Uniqe ID for the fatality}
  \item{EVENT_ID}{Event to which the fatality occurred}
  \item{FATALITY_TYPE}{
    \describe{
      \item{D}{Direct}
      \item{I}{Indirect}
    }
  }
  \item{FATALITY_DATE}{Date of the fatality in '\%m/\%d/\%Y \%H:\%M:\%S' format}
  \item{FATALITY_AGE}{Age of the victim}
  \item{FATALITY_SEX}{Sex of the victim}
  \item{FATALITY_LOCATION}{See \code{\link{dff_fatality_locations}}}
  \item{EVENT_YEARMONTH}{Year and month of the event in '\%Y\%m' format}
}}
\source{
\url{https://www.ncdc.noaa.gov/stormevents/details.jsp}
}
\usage{
s.fatalities
}
\description{
A dataset identifying fatalities associated with events in `details`.
}
\keyword{datasets}

