% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s.locations}
\alias{s.locations}
\title{Locations Dataset}
\format{A dataframe with 11 variables
\describe{
  \item{YEARMONTH}{Year and month of the event in '\%Y\%m' format}
  \item{EPISODE_ID}{A unique ID assigned by the NWS for a storm episode. Contains many EVENT_ID}
  \item{EVENT_ID}{A unique ID assigned by the NWS to a single event. Belongs to EPISODE_ID}
  \item{LOCATION_INDEX}{Number assigned by NWS to specific locations within the same event}
  \item{RANGE}{Distance in miles of event to location}
  \item{AZIMUTH}{Compass direction of event to location}
  \item{LOCATION}{Name of location}
  \item{LATITUDE}{Latitude of location}
  \item{LONGITUDE}{Longitude of location}
  \item{LAT2}{Unknown}
  \item{LON2}{Unknown}
}}
\source{
\url{https://www.ncdc.noaa.gov/stormevents/details.jsp}
}
\usage{
s.locations
}
\description{
A dataset identifying locations for events. I'm unsure for now the differences 
between this dataset and the location data in `details`.
}
\keyword{datasets}

