% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{timezone_fields}
\alias{timezone_fields}
\title{timezone_fields}
\usage{
timezone_fields()
}
\value{
character vector of required timezone variables
}
\description{
Required fields for timezone objects
}
\examples{
 \dontrun{
    if(all(names(DT)) \%in\% timezone_fields())...
    names(DT) <- timezone_fields()
  }
}

