% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{tz_to_dt}
\alias{tz_to_dt}
\title{Seperate timezones into data tables}
\usage{
tz_to_dt(DT = NULL)
}
\arguments{
\item{DT}{EVENT_ID and CZ_TIMEZONE field only}
}
\value{
Creates individual data tables for each unique timezone
}
\description{
Deprecated. Keeping for analysis of the crappy timezone data in the 
     dataset but provides no general use, I don't think.
}
\examples{
\dontrun{
  tz_to_dt(DT[, .(EVENT_ID, CZ_TIMEZONE)])
}
}

