% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details_damages}
\alias{details_damages}
\title{details_damages}
\usage{
details_damages(df)
}
\value{
df[,c("EVENT_ID", ["DAMAGE_PROPERTY"|"DAMAGE_CROPS"])]
}
\description{
Clean DAMAGE_PROPERTY or DAMAGE_CROPS
}
\details{
In each of these variables values are often represented like 2.5K or 
3M or 0.09K. The character key and numeric value are extracted then value is
multipled against k:
\itemize{
  \item B 1e+09
  \item M 1e+06
  \item K 1e+03
  \item H 1e+02
}
Numeric values replace the string values of the original variable. Values of 
0 and NA are removed.
}

