% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details_date_time}
\alias{details_date_time}
\title{details_date_time}
\usage{
details_date_time(df)
}
\arguments{
\item{df}{A dataframe with EVENT_ID and *_YEARMONTH, *_DAY and *_DATE_TIME 
where * is one of either BEGIN or END.}
}
\value{
A dataframe with EVENT_ID and *_DATE_TIME in in '%Y-%m-%d %H:%M:%S' format.
}
\description{
Reformat BEGIN_DATE_TIME or END_DATE_TIME to '%Y-%m-%d %T' format.
}
\details{
Takes one of details$BEGIN_DATE_TIME or details$END_DATE_TIME 
along with details$[BEGIN|END]_YEARMONTH, details$[BEGIN|END]_DAY 
and returns a clean, reformatted value of class character with no timezone. 
Times are assumed to be local times.
}

