% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{fips_dt}
\alias{fips_dt}
\title{Return FIPS dataset}
\usage{
fips_dt()
}
\value{
dataset of FIPS codes.
}
\description{
See \url{http://www.nws.noaa.gov/geodata/catalog/wsom/html/cntyzone.htm}
}
\examples{
\dontrun{
FIPS <- fips_dt()
}

}

