% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(year = NULL, ds = .datasets_available(), clean = TRUE)
}
\arguments{
\item{year}{A single number or numeric vector.}

\item{ds}{A string or character vector.}

\item{clean}{Clean and reorganize the dataset. TRUE by default.}
}
\value{
Returns a dataframe for each year and ds requested.
}
\description{
Download NCDC Storm Events datasets
}
\details{
This is a wrapper function for \code{\link{get_details}}, 
  \code{\link{get_fatalities}} and \code{\link{get_locations}}. This function 
  can load all requested datasets. Use \code{\link{get_listings}} first to 
  see what data is available and the file size of the datasets before 
  downloading datasets.
  
  Please see \href{https://www.ncdc.noaa.gov/stormevents/ftp.jsp}{Storm Events Database} 
  for latest code book and file naming convention.
}
\examples{
\dontrun{
  # Return the details dataset cleaned for 2015.
  get_data(year = 2015, ds = "details")
}

\dontrun{
  # Return the details, fatalities datasets for 2015 to 2016
  get_data(year = 2015:2016, ds = c("details", "fatalities"))
}

\dontrun{
  # Get all datasets for 2015
  get_data(year = 2015)
}

\dontrun{
  # details, fatalities and locations raw
  get_data(year = 2015, clean = FALSE)
}
}

