% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Download NCDC Storm Events datasets}
\usage{
get_data(year = NULL, type = c("details", "fatalities", "locations"))
}
\arguments{
\item{year}{numeric, vector}

\item{type}{character, vector}
}
\value{
Returns a data.table for each year and type requested.
}
\description{
Loads all requested datasets. Strongly recommended to use get_listings() 
     first to see what data is available and the file size of the datasets 
     beforeDownloading Datasets
}
\examples{
\dontrun{
get_data(year = c(1990:1995), type = c("details", "fatalities"))
}

}

