% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_datasets}
\alias{get_datasets}
\title{Get datasets}
\usage{
get_datasets(dt_names)
}
\arguments{
\item{dt_names}{list of Name from get_listings}
}
\value{
a data table of all years requested by Type
}
\description{
Get datasets
}

