% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{get_details}
\alias{get_details}
\title{get_details}
\usage{
get_details(year = NULL, clean = TRUE)
}
\arguments{
\item{year}{numeric vector with year(s) to retrieve data. Can be time-consuming.}

\item{clean}{clean the data if TRUE (default). Otherwise, return as-is.}
}
\description{
Download the details datasets for \code{year}.
}

