% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_listings}
\alias{get_listings}
\title{get_listings}
\usage{
get_listings()
}
\value{
a dataframe
}
\description{
Get listing of files from NCDC Storm Events
}
\details{
Get listing of all datasets off the NCDC Storm Events website. 
  The function returns a dataframe showing all StormEvents files listed on 
  the NCDC server.
  
  The data table contains the following variables:
  \itemize{
    \item Name: Filename of the dataset
    \item Modified: Modified date of the dataset
    \item Size: Size of the gzip file in bytes
    \item Dataset: One of c("details", "fatalities", "locations")
    \item Year: Year of dataset
  }
  
  The listings of datasets can be found at 
  \url{http://www1.ncdc.noaa.gov/pub/data/swdi/stormevents/csvfiles/}
}
\examples{
datasets <- get_listings()
}

